requirements.txt
accelerate>=0.33.0
bitsandbytes>=0.43.1
datasets>=3.0.1
huggingface-hub>=0.24.0
peft>=0.12.0
pyyaml>=6.0.1
safetensors>=0.4.5
sentencepiece>=0.2.0
transformers @ git+https://github.com/huggingface/transformers
trl>=0.9.6
unsloth[base] @ git+https://github.com/unslothai/unsloth.git
unsloth_zoo[base] @ git+https://github.com/unslothai/unsloth-zoo
torch>=2.3.0
triton>=3.0.0
